+++ src/gradia-api-config.ts
@@ -59,7 +59,8 @@ export const configDefaults = {
     "grid-columns-min":          3,     /*  grid: min side-by-side tiles           */
     "grid-gap-horizontal":       40,    /*  grid: horizontal gap between tiles     */
     "grid-gap-vertical":         20,    /*  grid: vertical gap between tiles       */
-    "grid-node-width-equal":     true   /*  grid: force equal width of node boxes  */
+    "grid-node-width-equal":     true,  /*  grid: force equal width of node boxes  */
+    "grid-node-height-equal":    true   /*  grid: force equal height of node boxes */
 }
 export type Config = typeof configDefaults